+++ qvarnmr/testing/utils.py
@@ -1,3 +1,9 @@
+from itertools import count
+
+from qvarnmr.handlers import get_handlers
+from qvarnmr.processor import get_changes, process_changes
+
+
 def cleaned(resource):
     return {k: v for k, v in resource.items() if k not in ('id', 'revision')}
 
@@ -41,3 +47,13 @@ def update_resource(qvarn, resource_type, resource_id, **kwargs):
     resource = qvarn.get(resource_type, resource_id)
     resource.update(kwargs)
     qvarn.update(resource_type, resource_id, resource)
+
+
+def process(qvarn, listeners, config, limit=10):
+    counter = count()
+    changes_processed = 1
+    mappers, reducers = get_handlers(config)
+    while changes_processed > 0:
+        changes = get_changes(qvarn, listeners)
+        changes_processed = process_changes(qvarn, config, changes, mappers, reducers)
+        assert next(counter) < limit