+++ tests/test_processor.py
@@ -1,9 +1,9 @@
 from qvarnmr.processor import UPDATED
-from qvarnmr.processor import process_map, process
+from qvarnmr.processor import process_map
 from qvarnmr.func import item, value
 from qvarnmr.handlers import get_handlers
 from qvarnmr.listeners import get_or_create_listeners
-from qvarnmr.testing.utils import get_mapped_data, get_resource_values, update_resource
+from qvarnmr.testing.utils import get_mapped_data, get_resource_values, update_resource, process
 
 
 SCHEMA = {
@@ -87,11 +87,10 @@ def test_process_map_resync_samever(pretender, qvarn):
         '_mr_version': 1,
     })
 
-    # Process map handlers in resync mode, since map handler version matches one already saved,
+    # Process map handlers in resync mode, since map handler version matches the one already saved,
     # nothing should be done.
     mappers, reducers = get_handlers(config)
-    assert list(process_map(qvarn, 'source', UPDATED, data['id'], mappers['source'],
-                            resync=True)) == []
+    assert process_map(qvarn, 'source', UPDATED, data['id'], mappers['source'], resync=True) == 0
 
     mapped = get_mapped_data(qvarn, 'map_target')
     assert mapped[data['id']] == {